+++ src/main/java/org/checkmate/user/controller/view/MainPageController.java
@@ -1,63 +1,73 @@
 package org.checkmate.user.controller.view;
 
-import javafx.animation.TranslateTransition;
-import javafx.event.ActionEvent;
-import javafx.fxml.FXML;
-import javafx.fxml.Initializable;
-import javafx.scene.control.Label;
-import javafx.scene.layout.AnchorPane;
-import javafx.util.Duration;
+import static org.checkmate.user.util.FilePath.READ_MY_INFO_FX;
+import static org.checkmate.user.util.FilePath.READ_MY_INFO_ST;
 
 import java.net.URL;
 import java.util.ResourceBundle;
-import org.checkmate.common.dto.response.LoginResponseDto;
+import javafx.application.Platform;
+import javafx.event.ActionEvent;
+import javafx.fxml.FXML;
+import javafx.fxml.Initializable;
+import javafx.scene.control.Hyperlink;
+
+import org.checkmate.common.controller.view.SceneManager;
 import org.checkmate.common.util.LoginSession;
 
 public class MainPageController implements Initializable {
 
-    @FXML private Label Menu;
-    @FXML private Label MenuBack;
-    @FXML private AnchorPane slider;
+    @FXML
+    private Hyperlink userNameLink;
+
+    @FXML
+    private void exit(ActionEvent event) {
+        Platform.exit();
+    }
+
+    /**
+     * 마이페이지 이동 이벤트 처리
+     */
+    @FXML
+    private void openMyPage(ActionEvent event) {
+        SceneManager sm = SceneManager.getInstance();
+        sm.moveScene(READ_MY_INFO_FX.getFilePath());
+    }
+
+    /**
+     * 도서 대여 이벤트 처리
+     */
+    @FXML
+    public void bookLoan(ActionEvent actionEvent) {
+        SceneManager sm = SceneManager.getInstance();
+        sm.moveScene("/org/checkmate/view/layouts/user/sidebarSamplePage.fxml");
+    }
+
+    // 대여정보
+    @FXML
+    public void loanStatus(ActionEvent actionEvent) {
+        SceneManager sm = SceneManager.getInstance();
+        sm.moveScene("/org/checkmate/view/layouts/user/sidebarSamplePage.fxml");
+    }
+
+    // 부서현황
+    @FXML
+    public void teamStatus(ActionEvent actionEvent) {
+        SceneManager sm = SceneManager.getInstance();
+        sm.moveScene("/org/checkmate/view/layouts/user/sidebarSamplePage.fxml");
+    }
+
+    // 도서신청
+    @FXML
+    public void reqPayment(ActionEvent actionEvent) {
+        SceneManager sm = SceneManager.getInstance();
+        sm.moveScene("/org/checkmate/view/layouts/user/sidebarSamplePage.fxml");
+    }
 
     @Override
     public void initialize(URL url, ResourceBundle resourceBundle) {
-        LoginSession loginSession = LoginSession.getInstance();
-        initializePage(loginSession.getMemberInfo());
+        var session = LoginSession.getInstance();
+        var memberInfo = session.getMemberInfo();
+        userNameLink.setText(memberInfo.getEName());
     }
 
-    private void initializePage(LoginResponseDto responseDto) {
-        slider.setTranslateX(-176);
-
-        Menu.setOnMouseClicked(event -> {
-            TranslateTransition slide = new TranslateTransition();
-            slide.setDuration(Duration.seconds(0.4));
-            slide.setNode(slider);
-
-            slide.setToX(0);
-            slide.play();
-
-            slider.setTranslateX(-176);
-
-            slide.setOnFinished((ActionEvent e) -> {
-                Menu.setVisible(false);
-                MenuBack.setVisible(true);
-            });
-        });
-
-        MenuBack.setOnMouseClicked(event -> {
-            TranslateTransition slide = new TranslateTransition();
-            slide.setDuration(Duration.seconds(0.4));
-            slide.setNode(slider);
-
-            slide.setToX(-176);
-            slide.play();
-
-            slider.setTranslateX(0);
-
-            slide.setOnFinished((ActionEvent e) -> {
-                Menu.setVisible(true);
-                MenuBack.setVisible(false);
-            });
-        });
-    }
 }
